+++ README.md
@@ -1,4 +1,4 @@
-# pymmcore-remote
+# pymmcore-plus
 
 Create & control [pymmcore](https://github.com/micro-manager/pymmcore) (python bindings for the C++ [micro-manager core](https://github.com/micro-manager/mmCoreAndDevices/))
 running in another process.
@@ -6,7 +6,7 @@ running in another process.
 
 
 ```python
-from pymmcore_remote import RemoteMMCore
+from pymmcore_plus import RemoteMMCore
 
 with RemoteMMCore() as mmcore:
     # 'demo' is a special option for the included CMMCorePlus
