+++ pymmcore_plus/_serialize.py
