+++ pyproject.toml
@@ -3,4 +3,4 @@ requires = ["setuptools>=42", "wheel", "setuptools_scm[toml]>=3.4"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools_scm]
-write_to = "pymmcore_remote/_version.py"
+write_to = "pymmcore_plus/_version.py"
