+++ setup.cfg
@@ -1,9 +1,9 @@
 [metadata]
-name = pymmcore-remote
+name = pymmcore-plus
 author = Talley Lambert
 author_email = talley.lambert@gmail.com
-url = https://github.com/tlambert03/pymmcore-remote
-download_url = https://github.com/tlambert03/pymmcore-remote
+url = https://github.com/tlambert03/pymmcore-plus
+download_url = https://github.com/tlambert03/pymmcore-plus
 license = BSD License (BSD-3)
 license_file = LICENSE
 description = Control pymmcore (micro-manager) in another process
@@ -11,8 +11,8 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 keywords = microscope, micro-manager, remote, multi-processing
 project_urls =
-    Source = https://github.com/tlambert03/pymmcore-remote
-    Tracker = https://github.com/tlambert03/pymmcore-remote/issues
+    Source = https://github.com/tlambert03/pymmcore-plus
+    Tracker = https://github.com/tlambert03/pymmcore-plus/issues
 classifiers =
     Development Status :: 3 - Alpha
     Intended Audience :: Developers
@@ -28,7 +28,7 @@ classifiers =
     Topic :: System :: Hardware :: Hardware Drivers
 
 [options]
-packages = pymmcore_remote
+packages = pymmcore_plus
 install_requires =
     pymmcore
     Pyro5>=5.10
@@ -58,7 +58,7 @@ addopts = -W error
 
 [isort]
 profile = black
-src_paths=pymmcore_remote
+src_paths=pymmcore_plus
 
 [flake8]
 exclude = _version.py,.eggs,examples
