+++ setup.py
@@ -1,6 +1,6 @@
 import setuptools
 
 setuptools.setup(
-    use_scm_version={"write_to": "pymmcore_remote/_version.py"},
+    use_scm_version={"write_to": "pymmcore_plus/_version.py"},
     setup_requires=["setuptools_scm"],
 )
